+++ CHANGES.rst
@@ -23,7 +23,7 @@ mast
 
 - Added ``Observations.download_file`` method to download a single file from MAST given an input
   data URI. [#1825]
-- Added case for passing a row to ``Observations.download_file` [#1881]
+- Added case for passing a row to ``Observations.download_file`` [#1881]
 - Removed deprecated ``Observations.get_hst_s3_uris()``, ``Observations.get_hst_s3_uri()``, 
   ``Core.get_token()``, ``Core.enable_s3_hst_dataset()``, ``Core.disable_s3_hst_dataset()`` and 
   variables obstype and silent [#1884]
@@ -99,6 +99,11 @@ gaia
 - Changed file names handling when downloading data. [#1784]
 - Improved code to handle bit data type. [#1784]
 - Prepared code to handle new datalink products. [#1784]
+  
+utils
+^^^^^
+
+- Read/write TableList [#1902]
 
 
 0.4.1 (2020-06-19)